# agentic_checkout/spec/openapi.agentic_checkout.yaml
openapi: 3.1.0
info:
  title: Agentic Checkout API
  version: "2026-01-30"
  description: |
    Merchant-implemented REST API for ChatGPT-driven checkout.
    Implements create, update (POST), retrieve (GET), complete, and cancel of checkout sessions.
servers:
  - url: https://merchant.example.com
security:
  - bearerAuth: []
...
paths:
  /checkout_sessions:
    post:
      tags: [CheckoutSessions]
      summary: Create a checkout session
      operationId: createCheckoutSession
      description: |
        Initializes a new checkout session from items and (optionally) buyer and fulfillment info.
        MUST return 201 with a rich, authoritative cart state.

        Agents MAY include `affiliate_attribution` with `touchpoint: "first"` to record first-touch attribution.
        This enables multi-touch attribution models where networks can capture both first and last touch.
      parameters:
        - $ref: "#/components/parameters/Authorization"
        - $ref: "#/components/parameters/ContentType"
        - $ref: "#/components/parameters/AcceptLanguage"
        - $ref: "#/components/parameters/UserAgent"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Timestamp"
        - $ref: "#/components/parameters/APIVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutSessionCreateRequest"
            examples:
              minimal:
                summary: Single item; no fulfillment details
                value:
                  items:
                    - id: "item_123"
                      quantity: 1
              with_address:
                summary: Single item with shipping address
                value:
                  items:
                    - id: "item_456"
                      quantity: 1
                  fulfillment_details:
                    name: "test"
                    phone_number: "15551234567"
                    email: "test@example.com"
                    address:
                      name: "test"
                      line_one: "555 Golden Gate Avenue"
                      line_two: "Apt 401"
                      city: "San Francisco"
                      state: "CA"
                      country: "US"
                      postal_code: "94102"
              with_first_touch_attribution:
                summary: With first-touch affiliate attribution
                value:
                  items:
                    - id: "item_123"
                      quantity: 1
                  affiliate_attribution:
                    provider: "impact.com"
                    token: "atp_01J8Z3WXYZ9ABC"
                    publisher_id: "pub_123"
                    touchpoint: "first"
      responses:
        "201":
          description: Created
          headers:
            Idempotency-Key:
              description: Echo of the request idempotency key
              schema: { type: string }
            Request-Id:
              description: Echo of the request correlation id
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "4XX":
          description: Client error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "5XX":
          description: Server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /checkout_sessions/{checkout_session_id}:
    parameters:
      - $ref: "#/components/parameters/Authorization"
      - $ref: "#/components/parameters/AcceptLanguage"
      - $ref: "#/components/parameters/UserAgent"
      - $ref: "#/components/parameters/IdempotencyKey"
      - $ref: "#/components/parameters/RequestId"
      - $ref: "#/components/parameters/Signature"
      - $ref: "#/components/parameters/Timestamp"
      - $ref: "#/components/parameters/APIVersion"
      - name: checkout_session_id
        in: path
        required: true
        description: "Unique identifier for the checkout session"
        schema: { type: string }
    post:
      tags: [CheckoutSessions]
      summary: Update a checkout session
      operationId: updateCheckoutSession
      description: |
        Applies changes (items, fulfillment address, fulfillment option) and returns an updated authoritative cart state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutSessionUpdateRequest"
            examples:
              set_fulfillment_option:
                summary: Select a fulfillment option
                value:
                  selected_fulfillment_options:
                    - type: "shipping"
                      shipping:
                        option_id: "fulfillment_option_456"
                        item_ids: ["item_123"]
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "4XX":
          description: Client error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "5XX":
          description: Server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    get:
      tags: [CheckoutSessions]
      summary: Retrieve a checkout session
      operationId: getCheckoutSession
      description: Returns the latest authoritative state for the checkout session.
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "404":
          description: Session not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /checkout_sessions/{checkout_session_id}/complete:
    parameters:
      - $ref: "#/components/parameters/Authorization"
      - $ref: "#/components/parameters/ContentType"
      - $ref: "#/components/parameters/AcceptLanguage"
      - $ref: "#/components/parameters/UserAgent"
      - $ref: "#/components/parameters/IdempotencyKey"
      - $ref: "#/components/parameters/RequestId"
      - $ref: "#/components/parameters/Signature"
      - $ref: "#/components/parameters/Timestamp"
      - $ref: "#/components/parameters/APIVersion"
      - name: checkout_session_id
        in: path
        required: true
        description: "Unique identifier for the checkout session"
        schema: { type: string }
    post:
      tags: [CheckoutSessions]
      summary: Complete a checkout session
      operationId: completeCheckoutSession
      description: |
        Finalizes the checkout by applying a payment method. MUST create an order and return completed state on success.

        Agents MAY include `affiliate_attribution` to ensure the merchant receives final attribution context.
        Attribution is stored alongside the resulting order but not returned in the response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutSessionCompleteRequest"
            examples:
              basic:
                summary: Complete with payment data
                value:
                  buyer:
                    first_name: "John"
                    last_name: "Smith"
                    email: "johnsmith@mail.com"
                  payment_data:
                    token: "spt_123"
                    provider: "stripe"
              with_affiliate_attribution:
                summary: Complete with last-touch affiliate attribution
                value:
                  buyer:
                    first_name: "John"
                    last_name: "Smith"
                    email: "johnsmith@mail.com"
                  payment_data:
                    token: "spt_123"
                    provider: "stripe"
                  affiliate_attribution:
                    provider: "impact.com"
                    token: "atp_01J8Z3WXYZ9ABC"
                    publisher_id: "pub_123"
                    touchpoint: "last"
      responses:
        "200":
          description: Completed session with order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionWithOrder"
        "4XX":
          description: Client error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "5XX":
          description: Server error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /checkout_sessions/{checkout_session_id}/cancel:
    parameters:
      - $ref: "#/components/parameters/Authorization"
      - $ref: "#/components/parameters/AcceptLanguage"
      - $ref: "#/components/parameters/UserAgent"
      - $ref: "#/components/parameters/IdempotencyKey"
      - $ref: "#/components/parameters/RequestId"
      - $ref: "#/components/parameters/Signature"
      - $ref: "#/components/parameters/Timestamp"
      - $ref: "#/components/parameters/APIVersion"
      - name: checkout_session_id
        in: path
        required: true
        description: "Unique identifier for the checkout session"
        schema: { type: string }
    post:
      tags: [CheckoutSessions]
      summary: Cancel a checkout session
      operationId: cancelCheckoutSession
      description: |
        Cancels a session if not already completed or canceled.
        Agents MAY include an intent_trace to communicate why the session is being abandoned.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelSessionRequest"
      responses:
        "200":
          description: Canceled session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
        "405":
          description: Not cancelable (already completed/canceled)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
  parameters:
    Authorization:
      name: Authorization
      in: header
      required: true
      description: "Bearer token for API authentication"
      schema: { type: string, example: "Bearer api_key_123" }
    ContentType:
      name: Content-Type
      in: header
      required: true
      description: "Must be application/json for requests with a body"
      schema: { type: string, example: "application/json" }
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: "Preferred language for response messages"
      schema: { type: string, example: "en-US" }
    UserAgent:
      name: User-Agent
      in: header
      required: false
      description: "Client application identifier"
      schema:
        {
          type: string,
          example: "ChatGPT/2.0 (Mac OS X 15.0.1; arm64; build 0)",
        }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: "Unique key to safely retry requests without duplicate processing"
      schema: { type: string, example: "idempotency_key_123" }
    RequestId:
      name: Request-Id
      in: header
      required: false
      description: "Unique identifier for request tracking and debugging"
      schema: { type: string, example: "request_id_123" }
    Signature:
      name: Signature
      in: header
      required: false
      description: "HMAC signature for webhook verification"
      schema: { type: string, example: "ZXltZX..." }
    Timestamp:
      name: Timestamp
      in: header
      required: false
      description: "RFC 3339 date-time string for request timing validation"
      schema:
        { type: string, format: date-time, example: "2026-01-30T10:30:00Z" }
    APIVersion:
      name: API-Version
      in: header
      required: true
      description: "API version date in YYYY-MM-DD format"
      schema: { type: string, example: "2026-01-30" }
  schemas:
...
    FulfillmentDetails:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, description: "Full name for fulfillment contact" }
        phone_number: { type: string, description: "Contact phone number" }
        email: { type: string, format: email, description: "Contact email address" }
        address: { $ref: "#/components/schemas/Address", description: "Fulfillment address" }
      example:
        name: "John Smith"
        phone_number: "15551234567"
        email: "john.smith@example.com"
        address:
          name: "John Smith"
          line_one: "555 Golden Gate Avenue"
          line_two: "Apt 401"
          city: "San Francisco"
          state: "CA"
          country: "US"
          postal_code: "94102"
...
    Buyer:
      type: object
      additionalProperties: false
      properties:
        first_name: { type: string, description: "Buyer's first name" }
        last_name: { type: string, description: "Buyer's last name" }
        full_name: { type: string, description: "Buyer's full name" }
        email: { type: string, format: email, description: "Buyer's email address" }
        phone_number: { type: string, description: "Buyer's phone number" }
        customer_id: { type: string, description: "Merchant's internal customer identifier" }
        account_type: { type: string, enum: [guest, registered, business], description: "Type of buyer account" }
        authentication_status:
          { type: string, enum: [authenticated, guest, requires_signin], description: "Buyer's authentication status" }
        company: { $ref: "#/components/schemas/CompanyInfo", description: "Company information for business buyers" }
        loyalty: { $ref: "#/components/schemas/LoyaltyInfo", description: "Loyalty program information" }
        tax_exemption: { $ref: "#/components/schemas/TaxExemption", description: "Tax exemption details" }
      required: [email]
      example:
        first_name: "Alice"
        last_name: "Johnson"
        email: "alice.johnson@example.com"
        phone_number: "15559876543"
        customer_id: "cust_abc123"
        account_type: "registered"
        authentication_status: "authenticated"
...
    Capabilities:
      type: object
      additionalProperties: false
      description: |
        Capabilities object used in requests and responses.
        Context determines the party: requests are from Agents, responses are from Sellers.
        Seller responses contain the intersection of supported interventions.
      properties:
        payment:
          $ref: "#/components/schemas/Payment"
        interventions:
          $ref: "#/components/schemas/InterventionCapabilities"
        extensions:
          oneOf:
            - type: array
              items:
                type: string
              uniqueItems: true
              description: Extensions the agent understands (request). Simple identifiers like 'discount'.
            - type: array
              items:
                $ref: "#/components/schemas/ExtensionDeclaration"
              uniqueItems: true
              description: Active extensions for this session (response). Objects with name, extends, schema, spec.
          description: |
            Extensions supported by the party.
            Requests: array of extension identifiers.
            Responses: array of extension declaration objects.
      example:
        payment:
          handlers:
            - id: "handler_1"
              name: "dev.acp.tokenized.card"
              version: "2026-01-30"
              spec: "https://example.com/handler-spec"
              requires_delegate_payment: true
              requires_pci_compliance: false
              psp: "stripe"
              config_schema: "https://example.com/config-schema"
              instrument_schemas: ["https://example.com/instrument-schema"]
              config: {}
        interventions:
          supported: ["3ds"]
        extensions: ["discount"]
...
    Item:
      type: object
      additionalProperties: false
      properties:
        id: { type: string, description: "Unique identifier for the item" }
        name: { type: string, description: "Display name of the item" }
        unit_amount: { type: integer, description: "Price per unit in minor currency units (e.g. 100 cents for $1.00 or 100 for ¥100)" }
      required: [id]
      example:
        id: "item_123"
        name: "Wireless Headphones"
        unit_amount: 7999
...
    FulfillmentGroup:
      type: object
      additionalProperties: false
      properties:
        id: { type: string, description: "Unique identifier for this fulfillment group" }
        item_ids:
          type: array
          items: { type: string }
          description: "List of line item IDs in this fulfillment group"
        destination_type:
          {
            type: string,
            enum: [shipping, pickup, local_delivery, digital],
            description: "Type of fulfillment for this group"
          }
        fulfillment_details: { $ref: "#/components/schemas/FulfillmentDetails", description: "Fulfillment contact and address details" }
        location_id: { type: string, description: "Location identifier for pickup or local delivery" }
        instructions: { type: string, description: "Special fulfillment instructions" }
      required: [id, item_ids, destination_type]
      example:
        id: "fg_001"
        item_ids: ["li_001", "li_002"]
        destination_type: "shipping"
        instructions: "Leave package at front door"
...
    DiscountsRequest:
      type: object
      additionalProperties: false
      description: Discount codes input for checkout create/update requests.
      properties:
        codes:
          type: array
          items: { type: string }
          description: Discount codes to apply. Case-insensitive. Replaces previously submitted codes. Send empty array to clear.
      example:
        codes: ["SAVE20", "FREESHIP"]
...
    CheckoutSession:
      allOf:
        - $ref: "#/components/schemas/CheckoutSessionBase"
    CheckoutSessionWithOrder:
      allOf:
        - $ref: "#/components/schemas/CheckoutSessionBase"
        - type: object
          properties:
            order: { $ref: "#/components/schemas/Order" }
          required: [order]
      example:
        id: "cs_abc123"
        protocol:
          version: "2026-01-30"
        capabilities:
          payment:
            handlers:
              - id: "handler_stripe_card"
                name: "dev.acp.tokenized.card"
                version: "2026-01-30"
                spec: "https://example.com/payment-handlers/tokenized-card"
                requires_delegate_payment: true
                requires_pci_compliance: false
                psp: "stripe"
                config_schema: "https://example.com/schemas/card-config.json"
                instrument_schemas: ["https://example.com/schemas/card-instrument.json"]
                config: {}
          interventions:
            supported: ["3ds"]
          extensions:
            - name: "discount@2026-01-30"
              extends: ["$.CheckoutSession.discounts"]
              schema: "https://example.com/extensions/discount/schema.json"
              spec: "https://example.com/extensions/discount/spec"
        status: "completed"
        currency: "usd"
        line_items:
          - id: "li_001"
            item:
              id: "item_123"
              name: "Wireless Headphones"
              unit_amount: 7999
            quantity: 2
            name: "Premium Wireless Headphones"
            unit_amount: 7999
            availability_status: "in_stock"
            totals:
              - type: "subtotal"
                display_text: "Subtotal"
                amount: 15998
        totals:
          - type: "subtotal"
            display_text: "Subtotal"
            amount: 15998
          - type: "tax"
            display_text: "Sales Tax"
            amount: 1400
          - type: "total"
            display_text: "Total"
            amount: 17398
        fulfillment_options:
          - type: "shipping"
            id: "ship_standard"
            title: "Standard Shipping"
            carrier: "USPS"
            totals:
              - type: "fulfillment"
                display_text: "Shipping"
                amount: 599
        messages: []
        links:
          - type: "return_policy"
            title: "Return Policy"
            url: "https://example.com/returns"
        created_at: "2026-02-11T10:00:00Z"
        updated_at: "2026-02-11T10:30:00Z"
        order:
          id: "ord_123456"
          checkout_session_id: "cs_abc123"
          order_number: "ORD-2026-001"
          permalink_url: "https://example.com/orders/ord_123456"
          status: "confirmed"
    CheckoutSessionCreateRequest:
      type: object
      additionalProperties: false
      properties:
        buyer: { $ref: "#/components/schemas/Buyer", description: "Buyer information" }
        line_items:
          type: array
          items: { $ref: "#/components/schemas/Item" }
          minItems: 1
          description: "Items to add to the checkout session"
        currency: { type: string, description: "ISO 4217 currency code" }
        fulfillment_details: { $ref: "#/components/schemas/FulfillmentDetails", description: "Fulfillment contact and address details" }
        capabilities: { $ref: "#/components/schemas/Capabilities", description: "Agent capabilities and supported features" }
        fulfillment_groups:
          type: array
          items: { $ref: "#/components/schemas/FulfillmentGroup" }
          description: "Grouping of items by fulfillment method"
        affiliate_attribution:
          { $ref: "#/components/schemas/AffiliateAttribution", description: "Affiliate attribution data for first-touch tracking" }
        coupons:
          type: array
          items: { type: string }
          description: "DEPRECATED: Use discounts.codes instead. Discount codes to apply."
        discounts:
          $ref: "#/components/schemas/DiscountsRequest"
          description: Discount codes to apply to the checkout session.
        locale: { type: string, description: "Locale code for localizing content (e.g., 'en-US')" }
        timezone: { type: string, description: "IANA timezone identifier (e.g., 'America/New_York')" }
        quote_id: { type: string, description: "Quote identifier if this session is based on a quote" }
        metadata:
          type: object
          additionalProperties: true
          description: "Arbitrary metadata for merchant use"
      required: [line_items, currency, capabilities]
      example:
        line_items:
          - id: "item_123"
            name: "Wireless Headphones"
            unit_amount: 7999
        currency: "usd"
        capabilities:
          payment:
            handlers:
              - id: "handler_stripe_card"
                name: "dev.acp.tokenized.card"
                version: "2026-01-30"
                spec: "https://example.com/payment-handlers/tokenized-card"
                requires_delegate_payment: true
                requires_pci_compliance: false
                psp: "stripe"
                config_schema: "https://example.com/schemas/card-config.json"
                instrument_schemas: ["https://example.com/schemas/card-instrument.json"]
                config: {}
          interventions:
            supported: ["3ds"]
          extensions: ["discount"]
    CheckoutSessionUpdateRequest:
      type: object
      additionalProperties: false
      properties:
        buyer: { $ref: "#/components/schemas/Buyer", description: "Updated buyer information" }
        line_items:
          type: array
          items: { $ref: "#/components/schemas/Item" }
          description: "Items to update in the checkout session"
        fulfillment_details: { $ref: "#/components/schemas/FulfillmentDetails", description: "Updated fulfillment contact and address" }
        fulfillment_groups:
          type: array
          items: { $ref: "#/components/schemas/FulfillmentGroup" }
          description: "Updated fulfillment groupings"
        selected_fulfillment_options:
          type: array
          items: { $ref: "#/components/schemas/SelectedFulfillmentOption" }
          description: "Fulfillment option selected by the buyer"
        coupons:
          type: array
          items: { type: string }
          description: "DEPRECATED: Use discounts.codes instead. Discount codes to apply."
        discounts:
          $ref: "#/components/schemas/DiscountsRequest"
          description: Discount codes to apply. Replaces previously submitted codes.
      example:
        selected_fulfillment_options:
          - type: "shipping"
            option_id: "ship_standard"
            item_ids: ["li_001"]
...
    CheckoutSessionCompleteRequest:
      type: object
      additionalProperties: false
      properties:
        buyer: { $ref: "#/components/schemas/Buyer", description: "Final buyer information" }
        payment_data: { $ref: "#/components/schemas/PaymentData", description: "Payment method and details" }
        authentication_result:
          { $ref: "#/components/schemas/AuthenticationResult", description: "Authentication result for 3DS flows" }
        affiliate_attribution:
          { $ref: "#/components/schemas/AffiliateAttribution", description: "Affiliate attribution data for last-touch tracking" }
        risk_signals: { $ref: "#/components/schemas/RiskSignals", description: "Risk and fraud signals" }
      required: [payment_data]
      example:
        buyer:
          first_name: "John"
          last_name: "Smith"
          email: "johnsmith@mail.com"
        payment_data:
          handler_id: "handler_stripe_card"
          instrument:
            type: "card"
            credential:
              type: "spt"
              token: "spt_123"
...
    CancelSessionRequest:
      type: object
      properties:
        intent_trace: { $ref: "#/components/schemas/IntentTrace" }
      example:
        intent_trace:
          reason_code: "price_sensitivity"
          trace_summary: "Customer mentioned the price was higher than expected"
    Error:
      type: object
      additionalProperties: false
      description: |
        Protocol-level error returned in 4xx/5xx responses when the server cannot
        return a valid CheckoutSession at all (e.g. malformed request or unexpected
        failure). Use Error—not MessageError—when there is no valid session state
        to return. type semantics: invalid_request — malformed request, missing
        required fields, or invalid JSON; request_not_idempotent — idempotency
        violation; processing_error — unexpected server-side failure;
        service_unavailable — temporary unavailability.
      properties:
        type:
          {
            type: string,
            enum:
              [
                invalid_request,
                request_not_idempotent,
                processing_error,
                service_unavailable,
              ],
            description: "Error type indicating the category of protocol-level error"
          }
        code: { type: string, description: "Implementation-defined error code" }
        message: { type: string, description: "Human-readable error message" }
        param: { type: string, description: "RFC 9535 JSONPath (optional)" }
      required: [type, code, message]
      example:
        type: "invalid_request"
        code: "missing_required_field"
        message: "The 'currency' field is required"
        param: "$.currency"